+++ pg_lake_iceberg/tests/pytests/test_polaris_conn.py
@@ -1,4 +1,5 @@
 from utils_pytest import *
+from utils_polaris import *
 import server_params
