+++ pg_lake_table/tests/conftest.py
@@ -4,91 +4,14 @@ import psycopg2
 import psycopg2.extras
 import pytest
 import shutil
-import subprocess
 import queue
 import threading
 import time
-from pathlib import Path
 from utils_pytest import *
 import server_params
 
 
 reduce_werkzeug_log_level()
-
-
-@pytest.fixture(scope="session")
-def server_state():
-    data = {"pgduck_server_started": False}
-    return data
-
-
-@pytest.fixture(scope="session")
-def pgduck_server(installcheck, server_state, s3):
-
-    if installcheck:
-        yield None
-    else:
-        if not server_state["pgduck_server_started"]:
-            server, output_queue, stderr_thread = setup_pgduck_server()
-
-            server_state["pgduck_server_started"] = True
-
-            yield server, output_queue, stderr_thread
-
-            terminate_server(server, stderr_thread)
-
-            server_state["pgduck_server_started"] = False
-        else:
-            yield None
-
-
-@pytest.fixture(scope="session")
-def postgres(installcheck, server_state, s3):
-    if installcheck:
-        # re-running installcheck might cause different results otherwise
-        remove_duckdb_cache()
-    else:
-        pgduck_started = server_state["pgduck_server_started"]
-
-        if not pgduck_started:
-            server, output_queue, stderr_thread = setup_pgduck_server()
-            server_state["pgduck_server_started"] = True
-
-        start_postgres(
-            server_params.PG_DIR, server_params.PG_USER, server_params.PG_PORT
-        )
-
-        polaris_server = start_polaris_server_in_background()
-
-    yield
-
-    if not installcheck:
-        stop_postgres(server_params.PG_DIR)
-
-        if os.path.isdir(server_params.PG_DIR + "/base/pgsql_tmp"):
-            assert len(os.listdir(server_params.PG_DIR + "/base/pgsql_tmp")) == 0
-
-        # we had to start ourselves
-        if not pgduck_started:
-            terminate_server(server, stderr_thread)
-            server_state["pgduck_server_started"] = False
-
-        polaris_server.terminate()
-        try:
-            polaris_server.wait(timeout=10)
-        except subprocess.TimeoutExpired:
-            polaris_server.kill()
-            polaris_server.wait(timeout=10)
-        polaris_pid = Path(server_params.POLARIS_PID_FILE)
-        if polaris_pid.exists():
-            os.kill(int(polaris_pid.read_text().strip()), signal.SIGTERM)
-
-
-@pytest.fixture(scope="module")
-def pg_conn(postgres, app_user):
-    conn = open_pg_conn(app_user)
-    yield conn
-    conn.close()
 
 
 @pytest.fixture(scope="module")
@@ -182,42 +105,6 @@ def pgduck_conn(postgres):
     conn.close()
 
 
-@pytest.fixture(scope="session")
-def _s3_internal():
-    """Creates a single Moto S3 instance and shares it across tests."""
-    client, server = create_mock_s3()  # Create both client & server
-    yield client, server  # Yield both as a tuple
-    server.stop()  # Stop the server at the end
-
-
-@pytest.fixture(scope="session")
-def s3(_s3_internal):
-    """Returns only the client so existing tests remain unchanged."""
-    client, _ = _s3_internal  # Extract only the client
-    return client  # No tuple, so existing tests continue working
-
-
-@pytest.fixture(scope="session")
-def s3_server(_s3_internal):
-    """Extracts the server from the shared S3 fixture."""
-    _, server = _s3_internal  # Get the server instance from `_s3_internal`
-    return server  # Return only the server
-
-
-@pytest.fixture(scope="session")
-def gcs():
-    client, server = create_mock_gcs()
-    yield client
-    server.stop()
-
-
-@pytest.fixture(scope="session")
-def azure():
-    client, server = create_mock_azure_blob_storage()
-    yield client
-    terminate_process(server)
-
-
 @pytest.fixture(scope="module")
 def test_s3_path(request, s3):
     return f"s3://{TEST_BUCKET}/{request.node.name}"
@@ -285,38 +172,6 @@ def create_pushdown_tables(s3, pg_conn, extension):
 
     run_command("DROP SCHEMA test_window_function_pushdown CASCADE;", pg_conn)
     pg_conn.commit()
-
-
-# when --installcheck is passed to pytests,
-# override the variables to point to the
-# official pgduck_server settings
-# this trick helps us to use the existing
-# pgduck_server
-@pytest.fixture(autouse=True, scope="session")
-def configure_server_params(request):
-    if request.config.getoption("--installcheck"):
-        server_params.PGDUCK_PORT = 5332
-        server_params.DUCKDB_DATABASE_FILE_PATH = "/tmp/duckdb.db"
-        server_params.PGDUCK_UNIX_DOMAIN_PATH = "/tmp"
-        server_params.PGDUCK_CACHE_DIR = "/tmp/cache"
-
-        # Access environment variables if exists
-        server_params.PG_DATABASE = os.getenv(
-            "PGDATABASE", "regression"
-        )  # 'postgres' or a default
-        server_params.PG_USER = os.getenv(
-            "PGUSER", "postgres"
-        )  # 'postgres' or a postgres
-        server_params.PG_PASSWORD = os.getenv(
-            "PGPASSWORD", "postgres"
-        )  # 'postgres' or a postgres
-        server_params.PG_PORT = os.getenv("PGPORT", "5432")  # '5432' or a default
-        server_params.PG_HOST = os.getenv(
-            "PGHOST", "localhost"
-        )  # 'localhost' or a default
-
-        # mostly relevant for CI
-        server_params.PG_DIR = "/tmp/pg_installcheck_tests"
 
 
 @pytest.fixture(scope="module")
