+++ pg_lake_table/tests/pytests/test_data_file_pruning.py
@@ -1,5 +1,6 @@
 import pytest
 import server_params
+from utils_polaris import *
 from utils_pytest import *
 import re
