+++ pg_lake_table/tests/pytests/test_modify_iceberg_rest_table.py
@@ -1,6 +1,7 @@
 import pytest
 import psycopg2
 from utils_pytest import *
+from utils_polaris import *
 from spark_utils import *
 import json
 import re
