+++ pg_lake_table/tests/pytests/test_partition_pruning.py
@@ -1,5 +1,6 @@
 import pytest
 from datetime import date
+from utils_polaris import *
 from utils_pytest import *
 from spark_utils import *
 import re
