+++ pg_lake_table/tests/pytests/test_polaris_catalog_writable.py
@@ -1,4 +1,5 @@
 from utils_pytest import *
+from utils_polaris import *
 import server_params
 from urllib.parse import quote
 from urllib.parse import urlencode
