+++ pgduck_server/tests/conftest.py
@@ -16,14 +16,11 @@ import server_params
 
 @pytest.fixture(scope="session")
 def pgduck_server(request, installcheck, configure_server_params):
-
-    if not installcheck:
-        server, output_queue, stderr_thread = setup_pgduck_server()
-
-    yield
-
-    if not installcheck:
-        terminate_server(server, stderr_thread)
+    if installcheck:
+        yield None
+    else:
+        setup_pgduck_server()
+        yield
 
 
 # when --installcheck is passed to pytests,
