+++ pgduck_server/tests/pytests/conftest.py
@@ -31,24 +31,3 @@ def pgduck_conn(pgduck_server):
     )
     yield conn
     conn.close()
-
-
-@pytest.fixture(scope="session")
-def s3():
-    client, server = create_mock_s3()
-    yield client
-    server.stop()
-
-
-@pytest.fixture(scope="session")
-def gcs():
-    client, server = create_mock_gcs()
-    yield client
-    server.stop()
-
-
-@pytest.fixture(scope="session")
-def azure():
-    client, process = create_mock_azure_blob_storage()
-    yield client
-    terminate_process(process)
